+++ src/DataProtection/DataProtection.Autorest/custom/Enums/Enums.cs
@@ -56,7 +56,8 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Support
         Daily,
         Weekly,
         Monthly,
-        Yearly
+        Yearly,
+        Default_OperationalStore
     }
 
     public enum TagName