+++ src/DataProtection/DataProtection.Autorest/custom/Helpers/PolicyHelpers.ps1
@@ -42,7 +42,7 @@ function GetTaggingPriority {
 	)
 
 	process{
-		$priorityMap = @{"Default"=99;"Daily"=25;"Weekly"=20;"Monthly"=15;"Yearly"=10}
+		$priorityMap = @{"Default"=99;"Default_OperationalStore"=99;"Daily"=25;"Weekly"=20;"Monthly"=15;"Yearly"=10}
 		return $priorityMap[$Name]
 	}
 }
@@ -94,5 +94,96 @@ function GetBackupFrequenceFromTimeInterval
 	{
 		$backupFrequencyMap = @{"D"="Daily";"H"="Hourly";"W"="Weekly"}
 		return "Backup" + $backupFrequencyMap[$RepeatingTimeInterval[0][-1].ToString()]
+	}
+}
+
+function ValidateRetentionRuleMatchesMappedStore
+{
+	[Microsoft.Azure.PowerShell.Cmdlets.DataProtection.DoNotExportAttribute()]
+	param(
+		[Parameter(Mandatory=$true)]
+		[ValidateNotNullOrEmpty()]
+		[System.String]
+		$Name,
+
+		[Parameter(Mandatory=$true)]
+		$DefaultRetentionMapping,
+
+		[Parameter(Mandatory=$true)]
+		[ValidateNotNullOrEmpty()]
+		$LifeCycles,
+
+		[Parameter(Mandatory=$true)]
+		[ValidateNotNullOrEmpty()]
+		[System.String]
+		$DatasourceType
+	)
+
+	process
+	{
+		$mappedDefaultNames = @($DefaultRetentionMapping.PSObject.Properties.Value)
+		if($mappedDefaultNames -contains $Name){
+			foreach($lc in $LifeCycles){
+				$store = $lc.SourceDataStoreType.ToString()
+				$expectedName = $null
+				if($DefaultRetentionMapping.PSObject.Properties.Name -contains $store){
+					$expectedName = $DefaultRetentionMapping.$store
+				}
+				if($null -ne $expectedName -and $expectedName -ne $Name){
+					$reservedStores = @($DefaultRetentionMapping.PSObject.Properties | Where-Object { $_.Value -eq $Name } | ForEach-Object { $_.Name })
+					$reservedFor = $reservedStores -join ", "
+					throw "Retention rule '$Name' is reserved for source store '$reservedFor' on datasource type $DatasourceType. For source store '$store' use -Name $expectedName. See Get-Help Edit-AzDataProtectionPolicyRetentionRuleClientObject -Examples."
+				}
+			}
+		}
+		return $mappedDefaultNames
+	}
+}
+
+function ValidateExclusiveSourceStoreAssignment
+{
+	[Microsoft.Azure.PowerShell.Cmdlets.DataProtection.DoNotExportAttribute()]
+	param(
+		[Parameter(Mandatory=$true)]
+		[ValidateNotNullOrEmpty()]
+		[System.String]
+		$Name,
+
+		[Parameter(Mandatory=$true)]
+		$Manifest,
+
+		[Parameter(Mandatory=$true)]
+		$DefaultRetentionMapping,
+
+		[Parameter(Mandatory=$true)]
+		[ValidateNotNullOrEmpty()]
+		$LifeCycles,
+
+		[Parameter(Mandatory=$true)]
+		[ValidateNotNullOrEmpty()]
+		[System.String]
+		$DatasourceType
+	)
+
+	process
+	{
+		$exclusiveStores = @()
+		if($null -ne $Manifest.policySettings -and ($Manifest.policySettings.PSObject.Properties.Name -contains "exclusiveSourceDataStores")){
+			$exclusiveStores = @($Manifest.policySettings.exclusiveSourceDataStores)
+		}
+		if($exclusiveStores.Count -gt 0){
+			foreach($lc in $LifeCycles){
+				$store = $lc.SourceDataStoreType.ToString()
+				if($exclusiveStores -contains $store){
+					$expectedName = $null
+					if($DefaultRetentionMapping.PSObject.Properties.Name -contains $store){
+						$expectedName = $DefaultRetentionMapping.$store
+					}
+					if($null -ne $expectedName -and $expectedName -ne $Name){
+						throw "Source store '$store' on datasource type $DatasourceType is exclusive: only the '$expectedName' retention rule may carry an $store lifecycle. Use -Name $expectedName instead of -Name $Name, or remove the $store lifecycle from -LifeCycles."
+					}
+				}
+			}
+		}
 	}
 }