+++ src/DataProtection/DataProtection.Autorest/test/New-AzDataProtectionBackupPolicy.Tests.ps1
@@ -275,40 +275,6 @@ Describe 'New-AzDataProtectionBackupPolicy' {
         $policy | Should be $null
     }
 
-    It 'BlobHardeningOperationalPolicy' {
-        $subId = $env.TestBlobHardeningScenario.SubscriptionId
-        $resourceGroupName = $env.TestBlobHardeningScenario.ResourceGroupName
-        $vaultName = $env.TestBlobHardeningScenario.VaultName
-        $operationalPolicyName = $env.TestBlobHardeningScenario.OperationalPolicyName        
-        
-        #Remove policy
-        Remove-AzDataProtectionBackupPolicy -Name $operationalPolicyName -ResourceGroupName $resourceGroupName -SubscriptionId $subId -VaultName $vaultName
-
-        # Create operational policy 
-        $defaultPol = Get-AzDataProtectionPolicyTemplate -DatasourceType AzureBlob
-
-        # Remove schedule 
-        Edit-AzDataProtectionPolicyTriggerClientObject -Policy $defaultPol -RemoveSchedule 
-
-        $lifeCycleOperationalTier = New-AzDataProtectionRetentionLifeCycleClientObject -SourceDataStore OperationalStore -SourceRetentionDurationType Days -SourceRetentionDurationCount 30 
- 
-        Edit-AzDataProtectionPolicyRetentionRuleClientObject -Policy $defaultPol -Name Default -LifeCycles $lifeCycleOperationalTier -IsDefault $true -OverwriteLifeCycle $true
-
-        $opPolicy = New-AzDataProtectionBackupPolicy -SubscriptionId $subId -ResourceGroupName $resourceGroupName -VaultName $vaultName -Name $operationalPolicyName -Policy $defaultPol 
-        
-        # get operational policy - verify name 
-        $opPolicy = Get-AzDataProtectionBackupPolicy -ResourceGroupName $resourceGroupName -VaultName $vaultName -SubscriptionId $subId -Name $operationalPolicyName
-
-        $opPolicy.Name | Should be $operationalPolicyName
-        $opPolicy.Property.PolicyRule.Count | Should be 1
-        $opPolicy.Property.PolicyRule[0].Lifecycle[0].SourceDataStoreType | Should be "OperationalStore"
-
-        #Remove policy
-        Remove-AzDataProtectionBackupPolicy -Name $operationalPolicyName -ResourceGroupName $resourceGroupName -SubscriptionId $subId -VaultName $vaultName
-        $opPolicy = Get-AzDataProtectionBackupPolicy -ResourceGroupName $resourceGroupName -VaultName $vaultName -SubscriptionId $subId | Where-Object { $_.Name -match $operationalPolicyName }
-        $opPolicy | Should be $null
-    }
-
     It 'BlobHardeningVaultedPolicy' {
         $subId = $env.TestBlobHardeningScenario.SubscriptionId
         $resourceGroupName = $env.TestBlobHardeningScenario.ResourceGroupName
@@ -356,7 +322,7 @@ Describe 'New-AzDataProtectionBackupPolicy' {
         $defaultPol = Get-AzDataProtectionPolicyTemplate -DatasourceType AzureBlob
 
         $lifeCycleOperationalTier = New-AzDataProtectionRetentionLifeCycleClientObject -SourceDataStore OperationalStore -SourceRetentionDurationType Days -SourceRetentionDurationCount 30
-        Edit-AzDataProtectionPolicyRetentionRuleClientObject -Policy $defaultPol -Name Default -LifeCycles $lifeCycleOperationalTier -IsDefault $true -OverwriteLifeCycle $false  
+        Edit-AzDataProtectionPolicyRetentionRuleClientObject -Policy $defaultPol -Name Default_OperationalStore -LifeCycles $lifeCycleOperationalTier -IsDefault $true
 
         # Weekly - 7W
         $lifeCycleVaultTierWeekly = New-AzDataProtectionRetentionLifeCycleClientObject -SourceDataStore VaultStore -SourceRetentionDurationType Weeks -SourceRetentionDurationCount 7
@@ -398,7 +364,7 @@ Describe 'New-AzDataProtectionBackupPolicy' {
         $operationalVaultedPolicy.Name | Should be $operationalVaultedPolicyName        
         $operationalVaultedPolicy.Property.PolicyRule[-1].Lifecycle[0].SourceDataStoreType | Should be "VaultStore"
         $operationalVaultedPolicy.Property.PolicyRule.Lifecycle.SourceDataStoreType -contains "OperationalStore" | Should be $true
-
+        
         #Remove policy
         Remove-AzDataProtectionBackupPolicy -Name $operationalVaultedPolicy -ResourceGroupName $resourceGroupName -SubscriptionId $subId -VaultName $vaultName
         $pol = Get-AzDataProtectionBackupPolicy -ResourceGroupName $resourceGroupName -VaultName $vaultName -SubscriptionId $subId | Where-Object { $_.Name -match $operationalVaultedPolicy }