+++ src/DataProtection/DataProtection.Autorest/test/utils.ps1
@@ -36,7 +36,7 @@ function setupEnv() {
     # as default. You could change them if needed.
     $env.SubscriptionId = (Get-AzContext).Subscription.Id
     $env.Tenant = (Get-AzContext).Tenant.Id
-    $env.RecordDate = (Get-Date -Year 2026 -Month 05 -Day 06 -Hour 16 -Minute 21 -Second 00).ToString('dd-MM-yyyy-h-m-s')
+    $env.RecordDate = (Get-Date -Year 2026 -Month 05 -Day 13 -Hour 23 -Minute 21 -Second 00).ToString('dd-MM-yyyy-h-m-s')
     # For any resources you created for test, you should add it to $env here.
 
     $BackupInstanceTestVariables = @{